fixed adding line/custom packs not working
faulty logic. the code expected "magane.available" storage item to have
already existed, and in the form of an empty array at the very least,
which obviously wouldn't be the case with new installations.

fixed favoriteStickersData always skipping the 2nd custom pack
typoed -1 with 1, the logic was meant to skip missing packs